openapi_codegen/pyproject.toml
[tool.mypy]
plugins = []
show_error_codes = true
strict = true

[[tool.mypy.overrides]]
module = [
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry]
name = "src"
version = "1.0.0"
description = ""
authors = ["kshramt"]
license = "AGPL"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.0"
pydantic = "^2.5.2"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
